refactor: 사이드바 탭 버튼 컴포넌트 수정 완료 (#8)
